CTQW model

Continuous-Time Quantum Walk

Full docs

Model functions

Quantum search constructor

Quantum walk constructor

Continuous-Time Quantum Walk

Exemplary implementation of continuous quantum walk model. The model is defined for arbitrary undirected graph. Hamiltonian is chosen to be adjacency matrix, and the evolution is made on the pure system of size equal to graph order. The definition can be found in Spatial search by quantum walk, written by Childs and Goldstone.

Full docs

Model functions

QSpatialSearch.initial_state — Method.

initial_state(qss::QSearch{AbstractCTQW})

Returns equal superposition of size size and type of qss.parameters[:hamiltonian].

julia> qss = QSearch(CTQW(CompleteGraph(4)), [1]);

julia> initial_state(qss)
4-element Array{Complex{Float64},1}:
 0.5+0.0im
 0.5+0.0im
 0.5+0.0im
 0.5+0.0im

source

QSpatialSearch.evolve — Method.

evolve(qss, state, runtime)

Returnes new state creates by evolving state by qss.parameters[:hamiltonian] for time runtime according to Schrödinger equation.

julia> qss = QSearch(CTQW(CompleteGraph(4)), [1]);

julia> evolve(qss, initial_state(qss), 1.)
4-element Array{Complex{Float64},1}:
 -0.128942+0.67431im
  0.219272+0.357976im
  0.219272+0.357976im
  0.219272+0.357976im

source

QSpatialSearch.measure — Method.

measure(::QWalkEvolution{<:AbstractCTQW}, state [, vertices])

Returns the probability of measuring each vertex from vertices from state according to AbstractCTQW model. If vertices is not provided, full measurement is made. For AbstractCTQW model measurement is done by taking square of absolute value of all elements of state.

julia> qss = QSearch(CTQW(CompleteGraph(4)), [1]);

julia> measure(qss, [sqrt(0.2), sqrt(0.3), sqrt(0.5)])
3-element Array{Float64,1}:
 0.2
 0.3
 0.5

 julia> measure(qss, [sqrt(0.2), sqrt(0.3), sqrt(0.5)], [2, 3])
 2-element Array{Float64,1}:
  0.3
  0.5

source

QSpatialSearch.matrix — Method.

matrix(ctqw::AbstractCTQW)

Returns the matrix symbol defining matrix graph used.

julia> matrix(CTQW(CompleteGraph(4)))
:adjacency

julia> matrix(CTQW(CompleteGraph(4), :laplacian))
:laplacian

source

Quantum search constructor

QSpatialSearch.QSearch — Method.

QSearch([type, ]ctqw::AbstractCTQW, marked [, jumpingrate, penalty])

Creates QSearch according to AbstractCTQW model. By default type equals Complex128, jumpingrate equals largest eigenvalue of adjacency matrix of graph if matrix(CTQW) outputs :adjacency and error otherwise, and penalty equals 0. By default hamiltonian is SparseMatrixCSC.

source

QSpatialSearch.check_qss — Function.

check_qss(qwalk, marked, parameters)

Checks whetver combination of qwalk, marked and parameters creates valid quantum search evolution. Note that whetver list of vertices marked are a subset of vertices of graph from qwalk is checked seperately in QSearch constructor.

Details concerning implementing check_qss for own models can be found in GitHub pages.

source

Quantum walk constructor

QSpatialSearch.QWalkSimulator — Type.

QWalkSimulator

source

QSpatialSearch.check_qwalksimulator — Function.

check_qwalksimulator(qwalk, parameters)

Checks whetver combination of qwalk and parameters creates valid quantum walk simulator.

Details concerning implementing check_qws for own models can be found in GitHub pages.

source