New model

QuantumWalk.initial_state — Function.

initial_state(qss)

Generates initial state for QWSearch qss. The type and value strongly depends on model of qss. For concrete model description please type ?initial_state(QWSearch{model}), for example for CTQW write ?initial_state(QWSearch{CTQW}).

Details concerning implementing initial_state for own models can be found in GitHub pages.

source

QuantumWalk.evolve — Function.

evolve(qwevolution{<:QWModelDiscr}, state)
evolve(qwevolution{<:QWModelCont}, state, time)

Evolve state according to qwevolution. time should be provided if model is continuous, otherwise one-step evolution should be performed. Type returned should be the same as type of state. The behaviour strongly depends on model of qwevolution. For concrete model description please type ?evolve(qwevolution{model}), for example for CTQW and QWSearch write ?evolve(QWSearch{CTQW}).

Details concerning implementing evolve for own models can be found in GitHub pages.

source

QuantumWalk.measure — Function.

measure(qwevolution, state[, vertices])

Measure state according to model from quantum walk evolution qwevolution. If vertices is provided, probabilities of given vertices should be provided. Otherwise full probability distribution should be provided. Output should be always of type Array{Float64}. The behaviour strongly depends on model of qwe. For concrete model and qwevolution description please type ?measure(qwevolution{model}), for example for CTQW and QWSearch write ?measure(QWSearch{CTQW}).

Details concerning implementing measure for own models can be found in GitHub pages.

source