Quantum walk evolution

Quantum evolution

The simplest quantum walk evolution. It simply takes the model and initial state from the user, makes and evolution and outputs the state or the probability distribution of measured state.

The dynamics requires evolve, measure and check_qwdynamics functions. It provides execute, execute_single, execute_single_measured, execute_all and execute_all_measured functions. Depending on the name it outputs single state or all states, measured or not measured. The execute combines the last four functions. In the case of type-stability requirement, we recommend to use the last four functions. execute_all and execute_all_measured are provided only for discrete quantum walk models.

Following functions are connected to the dynamics:

QuantumWalk.QWEvolution

QuantumWalk.check_qwdynamics

QuantumWalk.evolve

QuantumWalk.execute

QuantumWalk.execute_all

QuantumWalk.execute_all_measured

QuantumWalk.execute_single

QuantumWalk.execute_single_measured

QuantumWalk.measure

Example

julia> using LightGraphs, QuantumWalk

julia> qwe = QWEvolution(Szegedy(CompleteGraph(4)));

julia> state = rand(16); state = sparse(state/norm(state));

julia> execute_single(qwe, state, 4)
16-element SparseVector{Float64,Int64} with 16 stored entries:
  [1 ]  =  0.37243
  [2 ]  =  0.402212
  [3 ]  =  0.356914
  [4 ]  =  0.0384893
  [5 ]  =  0.230443
  [6 ]  =  0.0373444
  [7 ]  =  0.12092
  [8 ]  =  0.285178
  [9 ]  =  0.151519
  [10]  =  0.258755
  [11]  =  0.30374
  [12]  =  0.375624
  [13]  =  0.169741
  [14]  =  -0.0403198
  [15]  =  0.255573
  [16]  =  0.0344363

julia> execute_all_measured(qwe, state, 4)
4×5 Array{Float64,2}:
 0.273249  0.455568  0.308632  0.339199  0.429348
 0.20334   0.165956  0.188335  0.188466  0.150447
 0.244393  0.213254  0.378161  0.171268  0.323264
 0.279018  0.165221  0.124872  0.301066  0.0969412

julia> execute(qwe, state, 4, )
execute(qwd::QuantumWalk.QWDynamics, initstate, runtime::Real; all, measure) in QuantumWalk at /home/adam/.julia/v0.6/QuantumWalk/src/dynamics.jl:23
julia> execute(qwe, state, 4, measure=true, all=true)
4×5 Array{Float64,2}:
 0.273249  0.455568  0.308632  0.339199  0.429348
 0.20334   0.165956  0.188335  0.188466  0.150447
 0.244393  0.213254  0.378161  0.171268  0.323264
 0.279018  0.165221  0.124872  0.301066  0.0969412

Adjusting model to QWEvolution

Full docs

QuantumWalk.QWEvolution — Type.

QWEvolution(model, parameters)

Type describing standard quantum walk evolution. Needs implementation of

evolve(::QWEvolution{<:QWModelDiscr}, state) or evolve(::QWEvolution{<:QWModelCont}, state, time::Real)

measure(::QWEvolution, state)

check_qwevolution(::QWModelDiscr, parameters::Dict{Symbol})

proper constructors.

Offers functions

execute

execute_single

execute_single_measured

execute_all

execute_all_measured

source

QuantumWalk.check_qwdynamics — Method.

check_qwevolution(dynamictype, model, ..., parameters)

Checks whetver combination of the arguments creates valid quantum walk dynamics. The behaviour strongly depends on arguments.

source

QuantumWalk.evolve — Method.

evolve(qwevolution{<:QWModelDiscr}, state)
evolve(qwevolution{<:QWModelCont}, state, time)

Evolve state according to qwevolution. time should be provided if model is continuous, otherwise one-step evolution should be performed. Type returned should be the same as type of state. The behaviour strongly depends on arguments.

source

QuantumWalk.execute — Method.

execute(qwd, initstate, runtime[, all, measure])

Run proper execution function depending on given keywords. all and measure keywords defaults to false. For detailed description please see documentation of the corresponding function. Note that for all equal to true, the model specified by qwd needs to be disrete.

source

QuantumWalk.execute_all — Method.

execute_all(qwd::QWDynamics{<:QWModelDiscr}, initstate, runtime)

Evolve initstate acording to QWDynamics qwd for time runtime. Parameter runtime needs to be nonnegative and wqantum walk model needs to be discrete. Returns list of all states including the initstate and the final state.

julia> qwe = QWEvolution(Szegedy(CompleteGraph(4)));

julia> initstate = spzeros(16); initstate[1] = initstate[2] = 1/sqrt(2.);

julia> execute_all(qwe, initstate, 2)
3-element Array{SparseVector{Float64,Int64},1}:
   [1 ]  =  0.707107
  [2 ]  =  0.707107
   [1 ]  =  0.707107
  [2 ]  =  0.0785674
  [3 ]  =  -0.157135
  [4 ]  =  -0.157135
  [7 ]  =  0.31427
  [8 ]  =  0.31427
  [10]  =  -0.157135
  [12]  =  0.31427
  [14]  =  -0.157135
  [15]  =  0.31427
   [1 ]  =  0.707107
  [2 ]  =  0.427756
  [3 ]  =  -0.0698377
  [4 ]  =  -0.0698377
  [5 ]  =  -5.55112e-17
  [7 ]  =  -0.174594
  [8 ]  =  -0.174594
  [9 ]  =  0.31427
  [10]  =  -0.0698377
  [12]  =  0.139675
  [13]  =  0.31427
  [14]  =  -0.0698377
  [15]  =  0.139675

source

QuantumWalk.execute_all_measured — Method.

execute_all_measured(qwd::QWDynamics{<:QWModelDiscr}, initstate, runtime)

Evolve initstate acording to QWDynamics qwd for time runtime. Parameter runtime needs to be nonnegative and quantum walk model needs to be discrete. Returns matrix of type Matrix{Float64} for which i-th column is the probability distribution obtained from the measurement in i-1-th step.

julia> qwe = QWEvolution(Szegedy(CompleteGraph(4)));

julia> initstate = spzeros(16); initstate[1] = initstate[2] = 1/sqrt(2.);

julia> execute_all_measured(qwe, initstate, 2)
4×3 Array{Float64,2}:
 1.0  0.555556  0.69273
 0.0  0.197531  0.0609663
 0.0  0.123457  0.123152
 0.0  0.123457  0.123152

source

QuantumWalk.execute_single — Method.

execute_single(qwd, initstate, runtime)

Evolve initstate acording to QWDynamics qwd for time runtime. Parameter runtime needs to be nonnegative. If qwd subtype of QWModelDiscr, runtime needs to be of type Int.

julia> qwe = QWEvolution(Szegedy(CompleteGraph(4)));

julia> initstate = spzeros(16); initstate[1] = initstate[2] = 1/sqrt(2.);

julia> execute_single(qwe, initstate, 10)
16-element SparseVector{Float64,Int64} with 13 stored entries:
  [1 ]  =  0.707107
  [2 ]  =  0.6615
  [3 ]  =  -0.0114017
  [4 ]  =  -0.0114017
  [5 ]  =  -1.21431e-16
  [7 ]  =  -0.111417
  [8 ]  =  -0.111417
  [9 ]  =  0.13422
  [10]  =  -0.0114017
  [12]  =  0.0228034
  [13]  =  0.13422
  [14]  =  -0.0114017
  [15]  =  0.0228034

source

QuantumWalk.execute_single_measured — Method.

execute_single_measured(qwd, initstate, runtime)

Evolve initstate acording to QWDynamics qwd for time runtime and measure it in the end. Parameter runtime needs to be nonnegative. If qwd is subtype of QWModelDiscr, runtime needs to be Int.

julia> qwe = QWEvolution(Szegedy(CompleteGraph(4)));

julia> initstate = spzeros(16); initstate[1] = initstate[2] = 1/sqrt(2.);

julia> execute_single_measured(qwe, initstate, 10)
4-element Array{Float64,1}:
 0.937842
 0.0248275
 0.0186651
 0.0186651

source

QuantumWalk.measure — Method.

measure(qwevolution, state[, vertices])

Measure state according to model from quantum walk evolution qwevolution. If vertices is provided, probabilities of given vertices should be provided. Otherwise full probability distribution should be provided. Output should be always of type Array{Float64}. The behaviour strongly depends on arguments

source