Szegedy model

Quantum Search

Full docs

Types and its function

Continuous-Time Quantum Walk

Full docs

Model definition

General model functions

QWSearch documentation

QWEvolution documentation

Szegedy Quantum Walk

Full docs

Model definition

General model functions

Quantum search constructor

Quantum walk constructor

Quantum Search

Full docs

Types and its function

QSearchState

Quantum Walk models

Full docs

Types and its function

Quantum Dynamics

Full docs

Types and its function

Functions

Szegedy Quantum Walk

Full docs

Model definition

QuantumWalk.AbstractSzegedy — Type.

AbstractSzegedy

Abstract Szegedy model. Description of default can be found in https://arxiv.org/abs/1611.02238, where two oracle operator case is chosen. Default representation of AbstractSzegedy is Szegedy.

source

QuantumWalk.Szegedy — Type.

Szegedy(graph[, stochastic, checkstochastic])

Default representation of AbstractSzegedy. graph needs to be a subtype of AbstractGraph from LightGraphs module, stochastic needs to be a Real column stochastic matrix, checkstochastic is a flag which decides about checking the stochastic properties. stochastic defaults to uniform walk operator, checkstochastic deafults to false in case of default stochastic and true in case of user stochastic provided.

stochatsic is changed into sqrtstochastic by element-wise square root.

source

QuantumWalk.sqrtstochastic — Function.

sqrtstochastic(szegedy)

Returns the sqrtstochastic element of szegedy. After element-wise squaring a column-stochastic matrix is obtained.

julia> szegedy = Szegedy(CompleteGraph(4));

julia> sqrtstochastic(szegedy)
4×4 SparseMatrixCSC{Float64,Int64} with 12 stored entries:
  [2, 1]  =  0.57735
  [3, 1]  =  0.57735
  [4, 1]  =  0.57735
  [1, 2]  =  0.57735
  [3, 2]  =  0.57735
  [4, 2]  =  0.57735
  [1, 3]  =  0.57735
  [2, 3]  =  0.57735
  [4, 3]  =  0.57735
  [1, 4]  =  0.57735
  [2, 4]  =  0.57735
  [3, 4]  =  0.57735

julia> full(sqrtstochastic(szegedy).^2)
4×4 Array{Float64,2}:
 0.0       0.333333  0.333333  0.333333
 0.333333  0.0       0.333333  0.333333
 0.333333  0.333333  0.0       0.333333
 0.333333  0.333333  0.333333  0.0

source

General model functions

QuantumWalk.evolve — Method.

evolve(qwe::QWSearch{AbstractSzegedy}, state)

Multiplies state be each operator from operators from quatnum walk evolution qwe. Elements of operators and state should be of the same type.

julia> qss = QWSearch(Szegedy(CompleteGraph(4)), [1]);

julia> evolve(qss, initial_state(qss))
16-element SparseVector{Float64,Int64} with 12 stored entries:
  [2 ]  =  0.481125
  [3 ]  =  0.481125
  [4 ]  =  0.481125
  [5 ]  =  -0.288675
  [7 ]  =  -0.096225
  [8 ]  =  -0.096225
  [9 ]  =  -0.288675
  [10]  =  -0.096225
  [12]  =  -0.096225
  [13]  =  -0.288675
  [14]  =  -0.096225
  [15]  =  -0.096225

source

QuantumWalk.measure — Method.

measure(qwd::AbstractSzegedy, state[, vertices])

Performes a measurement on state on vertices. vertices defaults to list of all vertices. It is defined as the measurement of partially traced on second system state https://arxiv.org/abs/1611.02238.

julia> qss = QWSearch(Szegedy(CompleteGraph(4)), [1]);

julia> state = evolve(qss, initial_state(qss));

julia> measure(qss, state)
4-element Array{Float64,1}:
 0.694444
 0.101852
 0.101852
 0.101852

julia> measure(qss, state, [1,3])
2-element Array{Float64,1}:
 0.694444
 0.101852

source

Quantum search constructor

QuantumWalk.QWSearch — Method.

QWSearch(szegedy::AbstractSzegedy, marked [, penalty])

Creates QWSearch according to AbstractSzegedy model. penalty equals 0. It constructs evolution operators according to definition from https://arxiv.org/abs/1611.02238.

source

QuantumWalk.check_qwsearch — Method.

check_qwsearch(szegedy::AbstractSzegedy, marked, parameters)

Checks whetver combination of szegedy, marked and parameters produces valid QWSearch object. It checks where parameters consists of key :operators with corresponding value being list of SparseMatrixCSC. Furthermore operators needs to be square of size equals to square of graph(szegedy). order.

source

Quantum walk constructor

QuantumWalk.QWEvolution — Method.

QWEvolution(szegedy::AbstractSzegedy)

Creates QWEvolution according to AbstractSzegedy model. By default constructed operators is SparseMatrixCSC.

source

QuantumWalk.check_qwevolution — Method.

check_qwevolution(szegedy::AbstractSzegedy, marked, parameters)

Checks whetver combination of szegedy, marked and parameters produces valid QWEvolution object. It checks where parameters consists of key :operators with corresponding value being list of SparseMatrixCSC. Furthermore operators needs to be square of size equals to square of graph(szegedy). order.

source