Type hierarchy

Quantum Walk models

Full docs

Types and its function

QuantumWalk.QWModel — Type.

QWModel

Abstract supertype of all quantum walk models.

source

QuantumWalk.QWModelCont — Type.

QWModelCont

Abstract supertype of all continuous quantum walk models.

source

QuantumWalk.QWModelDiscr — Type.

QWModelDiscr

Abstract supertype of all discrete quantum walk models.

source

QuantumWalk.graph — Method.

graph(model)

Returns graph element of model.

julia> ctqw = CTQW(CompleteGraph(4))
QuantumWalk.CTQW({4, 6} undirected simple Int64 graph, :adjacency)

julia> graph(ctqw)
{4, 6} undirected simple Int64 graph

source

Quantum Dynamics

Full docs

Types and its function

QuantumWalk.QWDynamics — Type.

QWDynamics{T<:QWModel}

Abstract supertype of all dynamics on quantum walk models.

source

QuantumWalk.model — Function.

model(qwd)

Returns model element of qwd.

julia> qss = QWSearch(CTQW(CompleteGraph(4)), [1])
QuantumWalk.QWSearch{QuantumWalk.CTQW,Float64}(QuantumWalk.CTQW({4, 6} undirected simple Int64 graph, :adjacency), [1], Dict{Symbol,Any}(Pair{Symbol,Any}(:hamiltonian,
  [1, 1]  =  1.0+0.0im
  [2, 1]  =  0.333333+0.0im
  [3, 1]  =  0.333333+0.0im
  [4, 1]  =  0.333333+0.0im
  [1, 2]  =  0.333333+0.0im
  [3, 2]  =  0.333333+0.0im
  [4, 2]  =  0.333333+0.0im
  [1, 3]  =  0.333333+0.0im
  [2, 3]  =  0.333333+0.0im
  [4, 3]  =  0.333333+0.0im
  [1, 4]  =  0.333333+0.0im
  [2, 4]  =  0.333333+0.0im
  [3, 4]  =  0.333333+0.0im)), 0.0)

julia> model(qss)
QuantumWalk.CTQW({4, 6} undirected simple Int64 graph, :adjacency)

source

QuantumWalk.graph — Method.

graph(qwd)

Returns graph element of model from qwd. Equivalent to graph(model(qwd)).

julia> qss = QWSearch(CTQW(CompleteGraph(4)), [1])
QuantumWalk.QWSearch{QuantumWalk.CTQW,Float64}(QuantumWalk.CTQW({4, 6} undirected simple Int64 graph, :adjacency), [1], Dict{Symbol,Any}(Pair{Symbol,Any}(:hamiltonian,
  [1, 1]  =  1.0+0.0im
  [2, 1]  =  0.333333+0.0im
  [3, 1]  =  0.333333+0.0im
  [4, 1]  =  0.333333+0.0im
  [1, 2]  =  0.333333+0.0im
  [3, 2]  =  0.333333+0.0im
  [4, 2]  =  0.333333+0.0im
  [1, 3]  =  0.333333+0.0im
  [2, 3]  =  0.333333+0.0im
  [4, 3]  =  0.333333+0.0im
  [1, 4]  =  0.333333+0.0im
  [2, 4]  =  0.333333+0.0im
  [3, 4]  =  0.333333+0.0im)), 0.0)

julia> graph(qss)
{4, 6} undirected simple Int64 graph

source

QuantumWalk.parameters — Function.

parameters(qwd)

Returns parameters element of qwd.

julia> qss = QWSearch(CTQW(CompleteGraph(4)), [1])
QuantumWalk.QWSearch{QuantumWalk.CTQW,Float64}(QuantumWalk.CTQW({4, 6} undirected simple Int64 graph, :adjacency), [1], Dict{Symbol,Any}(Pair{Symbol,Any}(:hamiltonian,
  [1, 1]  =  1.0+0.0im
  [2, 1]  =  0.333333+0.0im
  [3, 1]  =  0.333333+0.0im
  [4, 1]  =  0.333333+0.0im
  [1, 2]  =  0.333333+0.0im
  [3, 2]  =  0.333333+0.0im
  [4, 2]  =  0.333333+0.0im
  [1, 3]  =  0.333333+0.0im
  [2, 3]  =  0.333333+0.0im
  [4, 3]  =  0.333333+0.0im
  [1, 4]  =  0.333333+0.0im
  [2, 4]  =  0.333333+0.0im
  [3, 4]  =  0.333333+0.0im)), 0.0)

julia> parameters(qss)
Dict{Symbol,Any} with 1 entry:
  :hamiltonian => …

source

Functions

QuantumWalk.execute — Method.

execute(qwd, initstate, runtime[, all, measure])

Run proper execution function depending on given keywords. all and measure keywords defaults to false. For detailed description please see documentation of corresponding function. Note that for all equal to true model in qwd needs to be disrete.

source

QuantumWalk.execute_single — Method.

execute_single(qwd, initstate, runtime)

Evolve initstate acording to QWDynamics qwd for time runtime. runtime needs to be nonnegative. If qwd is based on on QWModelDiscr, runtime needs to be Int.

julia> qwe = QWEvolution(Szegedy(CompleteGraph(4)));

julia> initstate = spzeros(16); initstate[1] = initstate[2] = 1/sqrt(2.);

julia> execute_single(qwe, initstate, 10)
16-element SparseVector{Float64,Int64} with 13 stored entries:
  [1 ]  =  0.707107
  [2 ]  =  0.6615
  [3 ]  =  -0.0114017
  [4 ]  =  -0.0114017
  [5 ]  =  -1.21431e-16
  [7 ]  =  -0.111417
  [8 ]  =  -0.111417
  [9 ]  =  0.13422
  [10]  =  -0.0114017
  [12]  =  0.0228034
  [13]  =  0.13422
  [14]  =  -0.0114017
  [15]  =  0.0228034

source

QuantumWalk.execute_single_measured — Method.

execute_single_measured(qwd, initstate, runtime)

Evolve initstate acording to QWDynamics qwd for time runtime. runtime needs to be nonnegative, and measure it in the end. If qwd is based on on QWModelDiscr, runtime needs to be Int.

julia> qwe = QWEvolution(Szegedy(CompleteGraph(4)));

julia> initstate = spzeros(16); initstate[1] = initstate[2] = 1/sqrt(2.);

julia> execute_single_measured(qwe, initstate, 10)
4-element Array{Float64,1}:
 0.937842
 0.0248275
 0.0186651
 0.0186651

source

QuantumWalk.execute_all — Method.

execute_all(qwd::QWDynamics{<:QWModelDiscr}, initstate, runtime)

Evolve initstate acording to QWDynamics qwd for time runtime. runtime needs to be nonnegative. Quantum walk model needs to be discrete. Returns list of all states including initstate and last state.

julia> qwe = QWEvolution(Szegedy(CompleteGraph(4)));

julia> initstate = spzeros(16); initstate[1] = initstate[2] = 1/sqrt(2.);

julia> execute_all(qwe, initstate, 2)
3-element Array{SparseVector{Float64,Int64},1}:
   [1 ]  =  0.707107
  [2 ]  =  0.707107
   [1 ]  =  0.707107
  [2 ]  =  0.0785674
  [3 ]  =  -0.157135
  [4 ]  =  -0.157135
  [7 ]  =  0.31427
  [8 ]  =  0.31427
  [10]  =  -0.157135
  [12]  =  0.31427
  [14]  =  -0.157135
  [15]  =  0.31427
   [1 ]  =  0.707107
  [2 ]  =  0.427756
  [3 ]  =  -0.0698377
  [4 ]  =  -0.0698377
  [5 ]  =  -5.55112e-17
  [7 ]  =  -0.174594
  [8 ]  =  -0.174594
  [9 ]  =  0.31427
  [10]  =  -0.0698377
  [12]  =  0.139675
  [13]  =  0.31427
  [14]  =  -0.0698377
  [15]  =  0.139675

source

QuantumWalk.execute_all_measured — Method.

execute_all_measured(qwd::QWDynamics{<:QWModelDiscr}, initstate, runtime)

Evolve initstate acording to QWDynamics qwd for time runtime. runtime needs to be nonnegative. Quantum walk model needs to be discrete. As a result return matrix of type Matrix{Float64} for which i-th column is measurement probability distribution in i-1-th step.

julia> qwe = QWEvolution(Szegedy(CompleteGraph(4)));

julia> initstate = spzeros(16); initstate[1] = initstate[2] = 1/sqrt(2.);

julia> execute_all_measured(qwe, initstate, 2)
4×3 Array{Float64,2}:
 1.0  0.555556  0.69273
 0.0  0.197531  0.0609663
 0.0  0.123457  0.123152
 0.0  0.123457  0.123152

source