CTQW model

Continuous-Time Quantum Walk

Implementation of continuous-time quantum walk. The model is defined for arbitrary undirected graph. Hamiltonian is chosen to be adjacency or Laplacian matrix. After the evolution, state is measured in canonic basis. The evolution is made on the pure system of size equal to graph order. The precise definition can be found in Spatial search by quantum walk by Childs and Goldstone, where both pure walk and search dynamics are described.

The abstract supertype is AbstractCTQW with its default realization CTQW. The model includes following types and methods:

QuantumWalk.AbstractCTQW

QuantumWalk.CTQW

QuantumWalk.QWEvolution

QuantumWalk.QWSearch

QuantumWalk.check_qwdynamics

QuantumWalk.check_qwdynamics

QuantumWalk.evolve

QuantumWalk.initial_state

QuantumWalk.matrix

QuantumWalk.measure

Full docs

QuantumWalk.AbstractCTQW — Type.

AbstractCTQW

Abstract CTQW model. By default evolve according to Schrödinger equation and performs measurmenet by taking square of absolute values of its elements. Default representation of AbstractCTQW is CTQW.

source

QuantumWalk.CTQW — Type.

CTQW(graph::Graph, matrix::Symbol)

Default representation of AbstractCTQW. matrix needs to be :adjacency or :laplacian and defaults to :adjacency.

source

QuantumWalk.QWEvolution — Method.

QWEvolution([type::Type{Number}, ]ctqw::AbstractCTQW)

Creates QWEvolution according to AbstractCTQW model. By default type equals Complex128. The hamiltonian is SparseMatrixCSC.

source

QuantumWalk.QWSearch — Method.

QWSearch([type::Type{T}, ]ctqw::AbstractCTQW, marked::Vector{Int}[, penalty::Real, jumpingrate::T]) where T<:Number

Creates QWSearch according to AbstractCTQW model. By default type equals Complex128, jumpingrate equals largest eigenvalue of adjacency matrix of graph if matrix(CTQW) outputs :adjacency and error otherwise, and penalty equals 0. The hamiltonian is SparseMatrixCSC.

QWSearch(qws::QWSearch{<:CTQW}; marked, penalty)

Updates quantum walk search to new subset of marked elements and new penalty. By default marked and penalty are the same as in qws.

source

QuantumWalk.check_qwdynamics — Method.

check_qwdynamics(QWSearch, ctqw::AbstractCTQW, parameters::Dict{Symbol}, marked::Vector{Int})

Checks whetver combination of ctqw, marked and parameters produces valid QWSearch object. It checks if parameters consists of key :hamiltonian with corresponding value being SparseMatrixCSC or Matrix. Furthermore the hamiltonian needs to be square of size equals to graph(ctqw) order. the hermiticity is not checked for efficiency issue.

source

QuantumWalk.check_qwdynamics — Method.

check_qwdynamics(QWEvolution, ctqw::AbstractCTQW, parameters::Dict{Symbol})

Checks iof combination of ctqw and parameters produces valid QWSearch object. It checks if parameters consists of key :hamiltonian with corresponding value being SparseMatrixCSC or Matrix. Furthermore the hamiltonian needs to be square of size equals to graph(ctqw) order. The hermiticity is not checked for efficiency issues.

source

QuantumWalk.evolve — Method.

evolve(qwd::QWDynamics{<:AbstractCTQW}, state::Vector{<:Number}, runtime::Real)

Returnes new state creates by evolving state by parameters(qwd)[:hamiltonian] for time runtime according to Schrödinger equation.

source

QuantumWalk.initial_state — Method.

initial_state(qws::QWSearch{AbstractCTQW})

Returns equal superposition of size size and type of parameters(qws)[:hamiltonian].

source

QuantumWalk.matrix — Method.

matrix(ctqw::AbstractCTQW)

Returns the matrix symbol defining matrix graph used.

source

QuantumWalk.measure — Method.

measure(::QWDynamics{<:AbstractCTQW}, state::Vector{<:Number}[, vertices::Vector{Int}])

Returns the probability of measuring each vertex from vertices from state according to AbstractCTQW model. If vertices is not provided, full measurement is made. The measurement is done by taking square of absolute value of all elements of state.

source