CTQW model

Continuous-Time Quantum Walk

Exemplary implementation of continuous quantum walk. The model is defined for arbitrary undirected graph. Hamiltonian is chosen to be adjacency or Laplacian matrix. The evolutions starts in equal superposition, and after time it is measured in canonic basis. The evolution is made on the pure system of size equal to graph order. The precise definition can be found in Spatial search by quantum walk, written by Childs and Goldstone, where both pure walk and search dynamics are described.

The abstract supertype is AbstractCTQW with its default realization CTQW. The model includes following types and methods:

QuantumWalk.AbstractCTQW

QuantumWalk.CTQW

QuantumWalk.QWEvolution

QuantumWalk.QWSearch

QuantumWalk.check_qwdynamics

QuantumWalk.check_qwdynamics

QuantumWalk.evolve

QuantumWalk.initial_state

QuantumWalk.matrix

QuantumWalk.measure

Full docs

QuantumWalk.AbstractCTQW — Type.

AbstractCTQW

Abstract CTQW model. By default evolve according to Schrödinger equation and performs measurmenet by taking square of absolute values of its elements. Default representation of AbstractCTQW is CTQW.

source

QuantumWalk.CTQW — Type.

CTQW(graph[, matrix])

Default representation of AbstractCTQW. graph needs to be of type Graph from LightGraphs module, matrix needs to be :adjacency or :laplacian and defaults to :adjacency.

source

QuantumWalk.QWEvolution — Method.

QWEvolution([type, ]ctqw::AbstractCTQW)

Creates QWEvolution according to AbstractCTQW model. By default type equals Complex128. By default constructed hamiltonian is SparseMatrixCSC.

source

QuantumWalk.QWSearch — Method.

QWSearch([type, ]ctqw::AbstractCTQW, marked [, jumpingrate, penalty])

Creates QWSearch according to AbstractCTQW model. By default type equals Complex128, jumpingrate equals largest eigenvalue of adjacency matrix of graph if matrix(CTQW) outputs :adjacency and error otherwise, and penalty equals 0. By default hamiltonian is SparseMatrixCSC.

source

QuantumWalk.check_qwdynamics — Method.

check_qwsearch(ctqw::AbstractCTQW, marked, parameters)

Checks whetver combination of ctqw, marked and parameters produces valid QWSearch object. It checks where parameters consists of key :hamiltonian with corresponding value being SparseMatrixCSC or Matrix. Furthermore the hamiltonian needs to be square of size equals to graph(ctqw) order.

source

QuantumWalk.check_qwdynamics — Method.

check_qwevolution(ctqw::AbstractCTQW, parameters)

Checks whetver combination of ctqw and parameters produces valid QWSearch object. It checks where parameters consists of key :hamiltonian with corresponding value being SparseMatrixCSC or Matrix. Furthermore the hamiltonian needs to be square of size equals to graph(ctqw) order.

source

QuantumWalk.evolve — Method.

evolve(qwd, state, runtime)

Returnes new state creates by evolving state by qwd.parameters[:hamiltonian] for time runtime according to Schrödinger equation.

julia> qss = QWSearch(CTQW(CompleteGraph(4)), [1]);

julia> evolve(qss, initial_state(qss), 1.)
4-element Array{Complex{Float64},1}:
 -0.128942+0.67431im
  0.219272+0.357976im
  0.219272+0.357976im
  0.219272+0.357976im

source

QuantumWalk.initial_state — Method.

initial_state(qss::QWSearch{AbstractCTQW})

Returns equal superposition of size size and type of qss.parameters[:hamiltonian].

julia> qss = QWSearch(CTQW(CompleteGraph(4)), [1]);

julia> initial_state(qss)
4-element Array{Complex{Float64},1}:
 0.5+0.0im
 0.5+0.0im
 0.5+0.0im
 0.5+0.0im

source

QuantumWalk.matrix — Method.

matrix(ctqw::AbstractCTQW)

Returns the matrix symbol defining matrix graph used.

julia> matrix(CTQW(CompleteGraph(4)))
:adjacency

julia> matrix(CTQW(CompleteGraph(4), :laplacian))
:laplacian

source

QuantumWalk.measure — Method.

measure(qwd::QWDynamics{<:AbstractCTQW}, state [, vertices])

Returns the probability of measuring each vertex from vertices from state according to AbstractCTQW model. If vertices is not provided, full measurement is made. For AbstractCTQW model measurement is done by taking square of absolute value of all elements of state.

julia> qss = QWSearch(CTQW(CompleteGraph(4)), [1]);

julia> measure(qss, [sqrt(0.2), sqrt(0.3), sqrt(0.5)])
3-element Array{Float64,1}:
 0.2
 0.3
 0.5

 julia> measure(qss, [sqrt(0.2), sqrt(0.3), sqrt(0.5)], [2, 3])
 2-element Array{Float64,1}:
  0.3
  0.5

source