Quantum search

Quantum Search

Quantum spatial search is an algorithm, which starts at some initial state (which depends on the graph structure), and runs for some time in order to cumulate amplitude at marked vertex. The algorithm is known to outperform classical search.

The dynamics requires evolve, measure, initial_state and check_qwdynamics functions. It provides execute, execute_single, execute_single_measured, execute_all and execute_all_measured functions - the description can be found in Quantum walk evolution section. The only difference is that QWSearch uses the state generated by initial_state function if not provided. Furthermore the function provides marked, penalty and maximize_quantum_search. The first two simple provides the parameters typical to QWSearch. The last one searches for optimal measure time. The maximization depends on the model (if it is continuous or discrete). While for discrete evolution obtaining reasonable result is guaranteed, it is not the case for the continuous one, as the optimization suffers for locating at local extrema.

The penalty is an additional time added in optimization. Note, that if we optimize time/success_probability(time) function, the optimum is always at time 0. This would imply that algorithm achieves full efficiency if it is instantly measured. This is misleading, as the time for constructing initial state and for measurement is ignored. Hence we need to include (usually small) additional time in penalty in order to get useful result. Note the time/success_probability(time) is at called expected runtime and can be obtained by expected_runtime function.

Some function as a result outputs QSearchState instead of the original state. It consists of the original state, the runtime and the probability of measuring each marked vertex. Those elements can be extracted by state, runtime and probability functions.

Following functions are connected to the quantum search:

QuantumWalk.QSearchState

QuantumWalk.QWSearch

QuantumWalk.check_qwdynamics

QuantumWalk.execute

QuantumWalk.execute_all

QuantumWalk.execute_all_measured

QuantumWalk.execute_single

QuantumWalk.execute_single

QuantumWalk.execute_single

QuantumWalk.execute_single_measured

QuantumWalk.expected_runtime

QuantumWalk.initial_state

QuantumWalk.marked

QuantumWalk.maximize_quantum_search

QuantumWalk.maximize_quantum_search

QuantumWalk.penalty

QuantumWalk.probability

QuantumWalk.runtime

QuantumWalk.state

Example

julia> n = 100;

julia> penalty_szegedy = log(n);

julia> qsearch = QWSearch(Szegedy(CompleteGraph(n)), [1], penalty_szegedy);

julia> runtime(maximize_quantum_search(qsearch))-penalty_szegedy
5.0

julia> probability(maximize_quantum_search(qsearch))
1-element Array{Float64,1}:
 0.569689

julia> execute_single_measured(qsearch, ceil(Int, pi*sqrt(100)/2))
100-element Array{Float64,1}:
 0.428475  
 0.00577298
 0.00577298
 0.00577298
 0.00577298
 0.00577298
 0.00577298
 0.00577298
 0.00577298
 ⋮         
 0.00577298
 0.00577298
 0.00577298
 0.00577298
 0.00577298
 0.00577298
 0.00577298
 0.00577298
 0.00577298

Adjusting model to QWSearch

Here we consider the example from the Quantum walk evolution section. We can consider random walk search as follows: at given step we check if we are at the marked vertex. If not, we continue evolution. Hence we need to cumulate the success probability at marked vertices. We can propose following implementation (including the functions from mentioned section). Again some additional assertion could be included for full funcionality.

##
function check_qwdynamics(::Type{QWSearch},
                          abs_stoch::UniformStochastic,
                          ::Array{Int},
                          parameters::Dict{Symbol,Any})
  @assert :stochastic ∈ keys(parameters) "parameters needs to have key stochastic"
  n = nv(graph(abs_stoch))
  @assert isa(parameters[:stochastic], SparseMatrixCSC{<:Real}) "value for :stochastic needs to be sparse matrix with real numbers"
  @assert size(parameters[:stochastic], 1) == size(parameters[:stochastic], 2) "Stochastic matrix needs to be square stochastic matrix"
  @assert mapslices(sum, parameters[:stochastic], 1)[1,:] ≈ ones(n) "Stochastic matrix needs to be square stochastic matrix of order graph"
end

function QWSearch(stoch::AbstractStochastic,
                  marked::Array{Int},
                  penalty::Real = 0.)
   parameters = Dict{Symbol,Any}(:stochastic => stochastic_matrix(graph(stoch)))

   QWSearch(stoch, marked, parameters, penalty)
end

function initial_state(qss::QWSearch{<:AbstractStochastic})
  n = nv(graph(qss))
  fill(1./n, n)
end

function evolve(qss::QWSearch{<:AbstractStochastic}, state::Vector{<:Real})
  old_probability = measure(qss, state, marked(qss))
  state[marked(qss)] = zero(marked(qss))
  state = stochastic_evolution(parameters(qss)[:stochastic], state)
  state[marked(qss)] += old_probability
  state
end

Note that for example measure function does not change. Below we provide an evolution simulation example.

julia> dynamic = QWSearch(UniformStochastic(CompleteGraph(100)), [1])
QuantumWalk.QWSearch{UniformStochastic{LightGraphs.SimpleGraphs.SimpleGraph{Int64}},Float64}(UniformStochastic{LightGraphs.SimpleGraphs.SimpleGraph{Int64}}({100, 4950} undirected simple Int64 graph), [1], Dict{Symbol,Any}(Pair{Symbol,Any}(:stochastic,
  [2  ,   1]  =  0.010101
  [3  ,   1]  =  0.010101
  [4  ,   1]  =  0.010101
  [5  ,   1]  =  0.010101
  [6  ,   1]  =  0.010101
  ⋮
  [94 , 100]  =  0.010101
  [95 , 100]  =  0.010101
  [96 , 100]  =  0.010101
  [97 , 100]  =  0.010101
  [98 , 100]  =  0.010101
  [99 , 100]  =  0.010101)), 0.0)

julia> measure(dynamic, execute_single(dynamic, 0), [1])
1-element Array{Float64,1}:
 0.01

julia> measure(dynamic, execute_single(dynamic, 40), [1])
1-element Array{Float64,1}:
 0.340416

julia> measure(dynamic, execute_single(dynamic, 1000), [1])
1-element Array{Float64,1}:
 0.999961

Full docs

QuantumWalk.QWSearch — Type.

QWSearch(model, marked, parameters, penalty)

Simulates quantum search on model with marked vertices and additional parameters. penalty represents the cost of initial state creation and measurement, which should be included for better optimization, see documentation of maximizing_function.

Needs implementation of

initial_state(::QWSearch)

evolve(::QWSearch{<:QWModelDiscr}, state) or evolve(::QWEvolution{<:QWModelCont}, state, time::Real)

measure(::QWSearch, state)

check_qwdynamics(::QWSearch, parameters::Dict{Symbol})

proper constructors.

Offers functions

execute

execute_single

execute_single_measured

execute_all

execute_all_measured

maximize_quantum_search

expected_runtime

and type QSearchState.

source

QuantumWalk.marked — Method.

marked(qss)

Returns marked element of qss.

julia> qss = QWSearch(CTQW(CompleteGraph(4)), [1]);

julia> marked(qss)
1-element Array{Int64,1}:
 1

source

QuantumWalk.penalty — Method.

penalty(qss)

Returns penalty element of qss.

julia> qss = QWSearch(CTQW(CompleteGraph(4)), [1]);

julia> penalty(qss)
0.0

source

QuantumWalk.check_qwdynamics — Method.

check_qwdynamics(model, marked, parameters)

Checks whetver combination of model, marked and parameters creates valid quantum search evolution. Note that whetver list of vertices marked are a subset of vertices of graph from model is checked seperately in QWSearch constructor.

Details concerning implementing check_qwdynamics for own models can be found in GitHub pages.

source

QuantumWalk.execute — Method.

execute(qss,[ initstate,] runtime[, all, measure])

Run proper execution of quantum spatial search depending on given keywords. The initial state is generated by initial_state(qss) function if not provided. all and measure keywords defaults to false. For detailed description please see documentation of corresponding function. Note that for all equal to true model in qss needs to be disrete.

source

QuantumWalk.execute_single — Method.

execute_single(qss,[ initstate,] runtime)

Evolve initstate acording to QWSearch qss for time runtime. The initial state is generated by initial_state(qss) function if not provided. runtime needs to be nonnegative. If qss is based on on QWModelDiscr, runtime needs to be Int. QSearchState{S}, where S is the state type, is returned.

julia> qss = QWSearch(Szegedy(CompleteGraph(4)), [1]);

julia> execute_single(qss, 4)
QuantumWalk.QSearchState{SparseVector{Float64,Int64},Int64}(  [2 ]  =  0.438623
  [3 ]  =  0.438623
  [4 ]  =  0.438623
  [5 ]  =  0.104937
  [7 ]  =  0.254884
  [8 ]  =  0.254884
  [9 ]  =  0.104937
  [10]  =  0.254884
  [12]  =  0.254884
  [13]  =  0.104937
  [14]  =  0.254884
  [15]  =  0.254884, [0.577169], 4)

source

QuantumWalk.execute_single — Method.

execute_single(qss,[ initstate,] runtime)

Evolve initstate acording to QWSearch qss for time runtime. The initial state is generated by initial_state(qss) function if not provided. runtime needs to be nonnegative. If qss is based on on QWModelDiscr, runtime needs to be Int. QSearchState{S}, where S is the state type, is returned.

julia> qss = QWSearch(Szegedy(CompleteGraph(4)), [1]);

julia> execute_single(qss, 4)
QuantumWalk.QSearchState{SparseVector{Float64,Int64},Int64}(  [2 ]  =  0.438623
  [3 ]  =  0.438623
  [4 ]  =  0.438623
  [5 ]  =  0.104937
  [7 ]  =  0.254884
  [8 ]  =  0.254884
  [9 ]  =  0.104937
  [10]  =  0.254884
  [12]  =  0.254884
  [13]  =  0.104937
  [14]  =  0.254884
  [15]  =  0.254884, [0.577169], 4)

source

QuantumWalk.execute_single — Method.

execute_single(qss,[ initstate,] runtime)

Evolve initstate acording to QWSearch qss for time runtime. The initial state is generated by initial_state(qss) function if not provided. runtime needs to be nonnegative. If qss is based on on QWModelDiscr, runtime needs to be Int. QSearchState{S}, where S is the state type, is returned.

julia> qss = QWSearch(Szegedy(CompleteGraph(4)), [1]);

julia> execute_single(qss, 4)
QuantumWalk.QSearchState{SparseVector{Float64,Int64},Int64}(  [2 ]  =  0.438623
  [3 ]  =  0.438623
  [4 ]  =  0.438623
  [5 ]  =  0.104937
  [7 ]  =  0.254884
  [8 ]  =  0.254884
  [9 ]  =  0.104937
  [10]  =  0.254884
  [12]  =  0.254884
  [13]  =  0.104937
  [14]  =  0.254884
  [15]  =  0.254884, [0.577169], 4)

source

QuantumWalk.execute_single_measured — Method.

execute_single_measured(qss, [ initstate,] runtime)

Evolve initstate acording to QWSearch qss for time runtime. The initial state is generated by initial_state(qss) function if not provided. runtime needs to be nonnegative. If qss is based on on QWModelDiscr, runtime needs to be Int. Probability measurement distribution is returned.

julia> qss = QWSearch(Szegedy(CompleteGraph(4)), [1]);

julia> execute_single_measured(qss, 4)
4-element Array{Float64,1}:
 0.577169
 0.140944
 0.140944
 0.140944

source

QuantumWalk.execute_all — Method.

execute_all(qss::QSWSearch{<:QWModelDiscr},[ initstate,] runtime)

Evolve initstate acording to QWSearch qss for time runtime. runtime needs to be nonnegative. The initial state is generated by initial_state(qss) function if not provided. Quantum walk model needs to be discrete. Returns list of all QSearchState{S} where S is state type including initstate and last state.

julia> qss = QWSearch(Szegedy(CompleteGraph(4)), [1]);

julia> execute_all(qss, 2)
3-element Array{QuantumWalk.QSearchState{SparseVector{Float64,Int64},Int64},1}:
 QuantumWalk.QSearchState{SparseVector{Float64,Int64},Int64}(  [2 ]  =  0.288675
  [3 ]  =  0.288675
  [4 ]  =  0.288675
  [5 ]  =  0.288675
  [7 ]  =  0.288675
  [8 ]  =  0.288675
  [9 ]  =  0.288675
  [10]  =  0.288675
  [12]  =  0.288675
  [13]  =  0.288675
  [14]  =  0.288675
  [15]  =  0.288675, [0.25], 0)
 QuantumWalk.QSearchState{SparseVector{Float64,Int64},Int64}(  [2 ]  =  0.481125
  [3 ]  =  0.481125
  [4 ]  =  0.481125
  [5 ]  =  -0.288675
  [7 ]  =  -0.096225
  [8 ]  =  -0.096225
  [9 ]  =  -0.288675
  [10]  =  -0.096225
  [12]  =  -0.096225
  [13]  =  -0.288675
  [14]  =  -0.096225
  [15]  =  -0.096225, [0.694444], 1)
 QuantumWalk.QSearchState{SparseVector{Float64,Int64},Int64}(  [2 ]  =  -0.138992
  [3 ]  =  -0.138992
  [4 ]  =  -0.138992
  [5 ]  =  0.032075
  [7 ]  =  -0.395592
  [8 ]  =  -0.395592
  [9 ]  =  0.032075
  [10]  =  -0.395592
  [12]  =  -0.395592
  [13]  =  0.032075
  [14]  =  -0.395592
  [15]  =  -0.395592, [0.0579561], 2)

source

QuantumWalk.execute_all_measured — Method.

execute_all_measured(qss::QWSearch{<:QWModelDiscr},[ initstate,] runtime)

Evolve initstate acording to QWSearch qss for time runtime. runtime needs to be nonnegative. The initial state is generated by initial_state(qss) function if not provided.Quantum walk model needs to be discrete. As a result return matrix of type Matrix{Float64} for which i-th column is measurement probability distribution in i-1-th step.

julia> qss = QWSearch(Szegedy(CompleteGraph(4)), [1]);

julia> execute_all_measured(qss, 2)
4×3 Array{Float64,2}:
 0.25  0.694444  0.0579561
 0.25  0.101852  0.314015
 0.25  0.101852  0.314015
 0.25  0.101852  0.314015

source

QuantumWalk.initial_state — Method.

initial_state(qss)

Generates initial state for QWSearch qss. The type and value strongly depends on model of qss. For concrete model description please type ?initial_state(QWSearch{model}), for example for CTQW write ?initial_state(QWSearch{CTQW}).

Details concerning implementing initial_state for own models can be found in GitHub pages.

source

QuantumWalk.maximize_quantum_search — Method.

maximize_quantum_search(qss::QWSearch{<:QWModelCont} [, maxtime, tstep])

Determines optimal runtime for continuous quantum walk models. The time is searched in [0, maxtime] interval, with penalty penalty(qss), which is added. It is recommended for penalty to be nonzero, otherwise time close to 0 is usually returned. Typically small penalty approximately equal to log(n) is enough, but optimal value may depend on the model or graph chosen.

The optimal time is chosen according to expected runtime, which equals to runtime over probability, which simulates the Bernoulli process based on QWModelCont.

tstep is used for primary grid search to search for determine intervale which is supsected to have small expected runtime. To large value may miss the optimal value, while to small may greatly increase runtime of the algorithm.

maxtime defaults to graph order n, tstep defaults to sqrt(n)/5. QSearchState is returned by deafult without penalty. Note that in general the probability is not maximal.

julia> qss = QWSearch(CTQW(CompleteGraph(100)), [1], 0.01, 1.);

julia> result = maximize_quantum_search(qss)
QuantumWalk.QSearchState{Array{Complex{Float64},1},Float64}(Complex{Float64}[0.621142+0.695665im, 0.0279736-0.023086im, 0.0279736-0.023086im, 0.0279736-0.023086im, 0.0279736-0.023086im, 0.0279736-0.023086im, 0.0279736-0.023086im, 0.0279736-0.023086im, 0.0279736-0.023086im, 0.0279736-0.023086im  …  0.0279736-0.023086im, 0.0279736-0.023086im, 0.0279736-0.023086im, 0.0279736-0.023086im, 0.0279736-0.023086im, 0.0279736-0.023086im, 0.0279736-0.023086im, 0.0279736-0.023086im, 0.0279736-0.023086im, 0.0279736-0.023086im], [0.869767], 12.99636940469214)

julia> expected_runtime(result)
14.94235723559316

julia> probability(result)
1-element Array{Float64,1}:
 0.869767

julia> probability(execute_single(qss, pi*sqrt(100)/2))
1-element Array{Float64,1}:
 1.0

source

QuantumWalk.maximize_quantum_search — Method.

maximize_quantum_search(qss::QWSearch{<:QWModelDiscr} [, runtime, mode])

Determines optimal runtime for continuous quantum walk models. The time is searched in [0, runtime] interval, with penalty penalty(qss), which is added. It is recommended for penalty to be nonzero, otherwise time close 0 is returned. Typically small penalty approximately equal to log(n) is enough, but optimal value may depend on the model or graph chosen.

The optimal time depende on chosen mode:

:firstmaxprob stops when probability start to decrease,

:firstmaxeff stops whene expected runtime start to increase,

:maxtimeeff chooses exhaustively the time from [0, runtime] with smallest expected time,

:maxtimeprob chooses exhaustively the time from [0, runtime] with maximal success probability,

:maxeff (default) finds optimal time with smallest expected time, usually faster than :maxtimefff.

Note last three modes always returns optimal time within the interval.

maxtime defaults to graph order n, mode defaults to :maxeff. QSearchState is returned by deafult without penalty.

julia> qss = QWSearch(Szegedy(CompleteGraph(200)), [1], 1);

julia> result = maximize_quantum_search(qss);

julia> runtime(result)
7

julia> probability(result)
1-element Array{Float64,1}:
 0.500016

julia> result = maximize_quantum_search(qss, 100, :maxtimeprob);

julia> runtime(result)
40

julia> probability(result)
1-element Array{Float64,1}:
 0.550938

source

QuantumWalk.QSearchState — Type.

QSearchState(state, probability, runtime)
QSearchState(qss, state, runtime)

Creates container which consists of state, success probability probability and running time runtime. Validity of probability and runtime is not checked.

In second case state is measured according to qss is made.

julia> qss = QWSearch(Szegedy(CompleteGraph(4)), [1]);

julia> result = QSearchState(qss, initial_state(qss), 0)
QuantumWalk.QSearchState{SparseVector{Float64,Int64},Int64}(  [2 ]  =  0.288675
  [3 ]  =  0.288675
  [4 ]  =  0.288675
  [5 ]  =  0.288675
  [7 ]  =  0.288675
  [8 ]  =  0.288675
  [9 ]  =  0.288675
  [10]  =  0.288675
  [12]  =  0.288675
  [13]  =  0.288675
  [14]  =  0.288675
  [15]  =  0.288675, [0.25], 0)

source

QuantumWalk.state — Function.

state(qsearchstate)

Returns the state of qsearchstate.

julia> qss = QWSearch(Szegedy(CompleteGraph(4)), [1]);

julia> result = QSearchState(qss, initial_state(qss), 0);

julia> state(result)
16-element SparseVector{Float64,Int64} with 12 stored entries:
  [2 ]  =  0.288675
  [3 ]  =  0.288675
  [4 ]  =  0.288675
  [5 ]  =  0.288675
  [7 ]  =  0.288675
  [8 ]  =  0.288675
  [9 ]  =  0.288675
  [10]  =  0.288675
  [12]  =  0.288675
  [13]  =  0.288675
  [14]  =  0.288675
  [15]  =  0.288675

source

QuantumWalk.probability — Function.

probability(qsearchstate)

Returns the list of probabilities of finding marked vertices.

julia> qss = QWSearch(Szegedy(CompleteGraph(4)), [1]);

julia> result = QSearchState(qss, initial_state(qss), 0);

julia> probability(result)
1-element Array{Float64,1}:
 0.25

source

QuantumWalk.runtime — Function.

runtime(qsearchstate)

Returns the time for which the state was calulated.

julia> qss = QWSearch(Szegedy(CompleteGraph(4)), [1]);

julia> result = QSearchState(qss, initial_state(qss), 0);

julia> runtime(result)
0

source

QuantumWalk.expected_runtime — Function.

expected_runtime(runtime, probability)
expected_runtime(state)

Returns the expected runtime needed for quantum walk, considering it as Bernoulli process. It equals to runtime/probability. In the case of state provided the measurement is made.

julia> qss = QWSearch(Szegedy(CompleteGraph(4)), [1]);

julia> result = execute(qss, 4);

julia> expected_runtime(result)
6.930377097077988

julia> expected_runtime(runtime(result), sum(probability(result)))
6.930377097077988

source