Quantum search

Quantum Search

Full docs

Types and its function

Continuous-Time Quantum Walk

Full docs

Model definition

General model functions

QWSearch documentation

QWEvolution documentation

Szegedy Quantum Walk

Full docs

Model definition

General model functions

Quantum search constructor

Quantum walk constructor

Quantum Search

Full docs

Types and its function

QSearchState

Quantum Walk models

Full docs

Types and its function

Quantum Dynamics

Full docs

Types and its function

Functions

Quantum Search

Full docs

Types and its function

QuantumWalk.QWSearch — Type.

QWSearch(model, marked, parameters, penalty)

Simulates quantum search on model with marked vertices and additional parameters. penalty represents the cost of initial state creation and measurement, which should be included for better optimization, see documentation of maximizing_function.

Needs implementation of

initial_state(::QWEvolution)

evolve(::QWEvolution{<:QWModelDiscr}, state) or evolve(::QWEvolution{<:QWModelCont}, state, time::Real)

measure(::QWEvolution, state)

check_qwsearch(::QWModelDiscr, parameters::Dict{Symbol})

proper constructors.

Offers functions

execute

execute_single

execute_single_measured

execute_all

execute_all_measured

maximize_quantum_search

expected_runtime

and type QSearchState.

source

QuantumWalk.marked — Function.

marked(qss)

Returns marked element of qss.

julia> qss = QWSearch(CTQW(CompleteGraph(4)), [1])
QuantumWalk.QWSearch{QuantumWalk.CTQW,Float64}(QuantumWalk.CTQW({4, 6} undirected simple Int64 graph, :adjacency), [1], Dict{Symbol,Any}(Pair{Symbol,Any}(:hamiltonian,
  [1, 1]  =  1.0+0.0im
  [2, 1]  =  0.333333+0.0im
  [3, 1]  =  0.333333+0.0im
  [4, 1]  =  0.333333+0.0im
  [1, 2]  =  0.333333+0.0im
  [3, 2]  =  0.333333+0.0im
  [4, 2]  =  0.333333+0.0im
  [1, 3]  =  0.333333+0.0im
  [2, 3]  =  0.333333+0.0im
  [4, 3]  =  0.333333+0.0im
  [1, 4]  =  0.333333+0.0im
  [2, 4]  =  0.333333+0.0im
  [3, 4]  =  0.333333+0.0im)), 0.0)

julia> marked(qss)
1-element Array{Int64,1}:
  1

source

QuantumWalk.penalty — Function.

penalty(qss)

Returns penalty element of qss.

julia> qss = QWSearch(CTQW(CompleteGraph(4)), [1])
QuantumWalk.QWSearch{QuantumWalk.CTQW,Float64}(QuantumWalk.CTQW({4, 6} undirected simple Int64 graph, :adjacency), [1], Dict{Symbol,Any}(Pair{Symbol,Any}(:hamiltonian,
  [1, 1]  =  1.0+0.0im
  [2, 1]  =  0.333333+0.0im
  [3, 1]  =  0.333333+0.0im
  [4, 1]  =  0.333333+0.0im
  [1, 2]  =  0.333333+0.0im
  [3, 2]  =  0.333333+0.0im
  [4, 2]  =  0.333333+0.0im
  [1, 3]  =  0.333333+0.0im
  [2, 3]  =  0.333333+0.0im
  [4, 3]  =  0.333333+0.0im
  [1, 4]  =  0.333333+0.0im
  [2, 4]  =  0.333333+0.0im
  [3, 4]  =  0.333333+0.0im)), 0.0)

julia> penalty(qss)
0.0

source

QuantumWalk.check_qwsearch — Function.

check_qwsearch(model, marked, parameters)

Checks whetver combination of model, marked and parameters creates valid quantum search evolution. Note that whetver list of vertices marked are a subset of vertices of graph from model is checked seperately in QWSearch constructor.

Details concerning implementing check_qwsearch for own models can be found in GitHub pages.

source

QuantumWalk.execute — Method.

execute(qss,[ initstate,] runtime[, all, measure])

Run proper execution of quantum spatial search depending on given keywords. The initial state is generated by initial_state(qss) function if not provided. all and measure keywords defaults to false. For detailed description please see documentation of corresponding function. Note that for all equal to true model in qss needs to be disrete.

source

QuantumWalk.execute_single — Method.

execute_single(qss,[ initstate,] runtime)

Evolve initstate acording to QWSearch qss for time runtime. The initial state is generated by initial_state(qss) function if not provided. runtime needs to be nonnegative. If qss is based on on QWModelDiscr, runtime needs to be Int. QSearchState{S}, where S is the state type, is returned.

julia> qss = QWSearch(Szegedy(CompleteGraph(4)), [1]);

julia> execute_single(qss, 4)
QuantumWalk.QSearchState{SparseVector{Float64,Int64},Int64}(  [2 ]  =  0.438623
  [3 ]  =  0.438623
  [4 ]  =  0.438623
  [5 ]  =  0.104937
  [7 ]  =  0.254884
  [8 ]  =  0.254884
  [9 ]  =  0.104937
  [10]  =  0.254884
  [12]  =  0.254884
  [13]  =  0.104937
  [14]  =  0.254884
  [15]  =  0.254884, [0.577169], 4)

source

QuantumWalk.execute_single_measured — Method.

execute_single_measured(qss, [ initstate,] runtime)

Evolve initstate acording to QWSearch qss for time runtime. The initial state is generated by initial_state(qss) function if not provided. runtime needs to be nonnegative. If qss is based on on QWModelDiscr, runtime needs to be Int. Probability measurement distribution is returned.

julia> qss = QWSearch(Szegedy(CompleteGraph(4)), [1]);

julia> execute_single_measured(qss, 4)
4-element Array{Float64,1}:
 0.577169
 0.140944
 0.140944
 0.140944

source

QuantumWalk.execute_all — Method.

execute_all(qss::QSWSearch{<:QWModelDiscr},[ initstate,] runtime)

Evolve initstate acording to QWSearch qss for time runtime. runtime needs to be nonnegative. The initial state is generated by initial_state(qss) function if not provided. Quantum walk model needs to be discrete. Returns list of all QSearchState{S} where S is state type including initstate and last state.

julia> qss = QWSearch(Szegedy(CompleteGraph(4)), [1]);

julia> execute_all(qss, 2)
3-element Array{QuantumWalk.QSearchState{SparseVector{Float64,Int64},Int64},1}:
 QuantumWalk.QSearchState{SparseVector{Float64,Int64},Int64}(  [2 ]  =  0.288675
  [3 ]  =  0.288675
  [4 ]  =  0.288675
  [5 ]  =  0.288675
  [7 ]  =  0.288675
  [8 ]  =  0.288675
  [9 ]  =  0.288675
  [10]  =  0.288675
  [12]  =  0.288675
  [13]  =  0.288675
  [14]  =  0.288675
  [15]  =  0.288675, [0.25], 0)
 QuantumWalk.QSearchState{SparseVector{Float64,Int64},Int64}(  [2 ]  =  0.481125
  [3 ]  =  0.481125
  [4 ]  =  0.481125
  [5 ]  =  -0.288675
  [7 ]  =  -0.096225
  [8 ]  =  -0.096225
  [9 ]  =  -0.288675
  [10]  =  -0.096225
  [12]  =  -0.096225
  [13]  =  -0.288675
  [14]  =  -0.096225
  [15]  =  -0.096225, [0.694444], 1)
 QuantumWalk.QSearchState{SparseVector{Float64,Int64},Int64}(  [2 ]  =  -0.138992
  [3 ]  =  -0.138992
  [4 ]  =  -0.138992
  [5 ]  =  0.032075
  [7 ]  =  -0.395592
  [8 ]  =  -0.395592
  [9 ]  =  0.032075
  [10]  =  -0.395592
  [12]  =  -0.395592
  [13]  =  0.032075
  [14]  =  -0.395592
  [15]  =  -0.395592, [0.0579561], 2)

source

QuantumWalk.execute_all_measured — Method.

execute_all_measured(qss::QWSearch{<:QWModelDiscr},[ initstate,] runtime)

Evolve initstate acording to QWSearch qss for time runtime. runtime needs to be nonnegative. The initial state is generated by initial_state(qss) function if not provided.Quantum walk model needs to be discrete. As a result return matrix of type Matrix{Float64} for which i-th column is measurement probability distribution in i-1-th step.

julia> qss = QWSearch(Szegedy(CompleteGraph(4)), [1]);

julia> execute_all_measured(qss, 2)
4×3 Array{Float64,2}:
 0.25  0.694444  0.0579561
 0.25  0.101852  0.314015
 0.25  0.101852  0.314015
 0.25  0.101852  0.314015

source

QuantumWalk.maximize_quantum_search — Function.

maximize_quantum_search(qss::QWSearch{<:QWModelCont} [, maxtime, tstep])

Determines optimal runtime for continuous quantum walk models. The time is searched in [0, maxtime] interval, with penalty penalty(qss), which is added. It is recommended for penalty to be nonzero, otherwise time close to 0 is usually returned. Typically small penalty approximately equal to log(n) is enough, but optimal value may depend on the model or graph chosen.

The optimal time is chosen according to expected runtime, which equals to runtime over probability, which simulates the Bernoulli process based on QWModelCont.

tstep is used for primary grid search to search for determine intervale which is supsected to have small expected runtime. To large value may miss the optimal value, while to small may greatly increase runtime of the algorithm.

maxtime defaults to graph order n, tstep defaults to sqrt(n)/5. QSearchState is returned by deafult without penalty. Note that in general the probability is not maximal

julia> using QuantumWalk,LightGraphs

julia> qss = QWSearch(CTQW(CompleteGraph(100)), [1], 0.01, 1.);

julia> result = maximize_quantum_search(qss)
QuantumWalk.QSearchState{Array{Complex{Float64},1},Float64}(Complex{Float64}[0.621142+0.695665im, 0.0279736-0.023086im, 0.0279736-0.023086im, 0.0279736-0.023086im, 0.0279736-0.023086im, 0.0279736-0.023086im, 0.0279736-0.023086im, 0.0279736-0.023086im, 0.0279736-0.023086im, 0.0279736-0.023086im  …  0.0279736-0.023086im, 0.0279736-0.023086im, 0.0279736-0.023086im, 0.0279736-0.023086im, 0.0279736-0.023086im, 0.0279736-0.023086im, 0.0279736-0.023086im, 0.0279736-0.023086im, 0.0279736-0.023086im, 0.0279736-0.023086im], [0.869767], 12.99636940469214)

julia> expected_runtime(result)
14.94235723559316

julia> probability(result)
1-element Array{Float64,1}:
 0.869767

julia> probability(execute_single(qss, pi*sqrt(100)/2))
1-element Array{Float64,1}:
 1.0

source

maximize_quantum_search(qss::QWSearch{<:QWModelDiscr} [, runtime, mode])

Determines optimal runtime for continuous quantum walk models. The time is searched in [0, runtime] interval, with penalty penalty(qss), which is added. It is recommended for penalty to be nonzero, otherwise time close 0 is returned. Typically small penalty approximately equal to log(n) is enough, but optimal value may depend on the model or graph chosen.

The optimal time depende on chosen mode:

:firstmaxprob stops when probability start to decrease,

:firstmaxeff stops whene expected runtime start to increase,

:maxtimeeff chooses exhaustively the time from [0, runtime] with smallest expected time,

:maxtimeprob chooses exhaustively the time from [0, runtime] with maximal success probability,

:maxeff (default) finds optimal time with smallest expected time, usually faster

than :maxtimefff.

Note last three modes always returns optimal time within the interval.

maxtime defaults to graph order n, mode defaults to :maxeff. QSearchState is returned by deafult without penalty.

julia> qss = QWSearch(Szegedy(CompleteGraph(200)), [1], 1);

julia> result = maximize_quantum_search(qss);

julia> runtime(result)
7

julia> probability(result)
1-element Array{Float64,1}:
 0.500016

julia> result = maximize_quantum_search(qss, 100, :maxtimeprob);

julia> runtime(result)
40

julia> probability(result)
1-element Array{Float64,1}:
 0.550938

source

QSearchState

QuantumWalk.QSearchState — Type.

QSearchState(state, probability, runtime)
QSearchState(qss, state, runtime)

Creates container which consists of state, success probability probability and running time runtime. Validity of probability and runtime is not checked.

In second case state is measured according to qss is made.

julia?> 2+2
5

julia> qss = QWSearch(Szegedy(CompleteGraph(4)), [1]);

julia> result = QSearchState(qss, initial_state(qss), 0)
QuantumWalk.QSearchState{SparseVector{Float64,Int64},Int64}(  [2 ]  =  0.288675
  [3 ]  =  0.288675
  [4 ]  =  0.288675
  [5 ]  =  0.288675
  [7 ]  =  0.288675
  [8 ]  =  0.288675
  [9 ]  =  0.288675
  [10]  =  0.288675
  [12]  =  0.288675
  [13]  =  0.288675
  [14]  =  0.288675
  [15]  =  0.288675, [0.25], 0)

source

QuantumWalk.state — Function.

state(qsearchstate)

Returns the state of qsearchstate.

julia> qss = QWSearch(Szegedy(CompleteGraph(4)), [1]);

julia> result = QSearchState(qss, initial_state(qss), 0);

julia> state(result)
16-element SparseVector{Float64,Int64} with 12 stored entries:
  [2 ]  =  0.288675
  [3 ]  =  0.288675
  [4 ]  =  0.288675
  [5 ]  =  0.288675
  [7 ]  =  0.288675
  [8 ]  =  0.288675
  [9 ]  =  0.288675
  [10]  =  0.288675
  [12]  =  0.288675
  [13]  =  0.288675
  [14]  =  0.288675
  [15]  =  0.288675

source

QuantumWalk.probability — Function.

probability(qsearchstate)

Returns the list of probabilities of finding marked vertices.

julia> qss = QWSearch(Szegedy(CompleteGraph(4)), [1]);

julia> result = QSearchState(qss, initial_state(qss), 0);

julia> probability(result)
1-element Array{Float64,1}:
 0.25

source

QuantumWalk.runtime — Function.

runtime(qsearchstate)

Returns the time for which the state was calulated.

julia> qss = QWSearch(Szegedy(CompleteGraph(4)), [1]);

julia> result = QSearchState(qss, initial_state(qss), 0);

julia> runtime(result)
0

source

QuantumWalk.expected_runtime — Function.

expected_runtime(runtime, probability)
expected_runtime(state)

Returns the expected runtime needed for quantum walk, considering it as Bernoulli process. It equals to runtime/probability. In the case of state provided the measurement is made.

julia> qss = QWSearch(Szegedy(CompleteGraph(4)), [1]);

julia> result = execute(qss, 4);

julia> expected_runtime(result)
6.930377097077988

julia> expected_runtime(runtime(result), sum(probability(result)))
6.930377097077988

source