Quantum walk simulator

Quantum Search

Full docs

Types and its function

QuantumWalk.QWEvolution — Type.

QWEvolution(model, parameters)

Type describing standard quantum walk evolution. Needs implementation of

evolve(::QWEvolution{<:QWModelDiscr}, state) or evolve(::QWEvolution{<:QWModelCont}, state, time::Real)

measure(::QWEvolution, state)

check_qwevolution(::QWModelDiscr, parameters::Dict{Symbol})

proper constructors.

Offers functions

execute

execute_single

execute_single_measured

execute_all

execute_all_measured

source

QuantumWalk.check_qwevolution — Function.

check_qwevolution(model, parameters)

Checks whetver combination of model and parameters creates valid quantum walk simulator.

Details concerning implementing check_qws for own models can be found in GitHub pages.

source