Szegedy model

Szegedy Quantum Walk

The Szegedy quantum walk is one of the most popular discrete quantum walk models. It takes stochastic matrix, turns it into unitary operator and use it for evolution. The evolution is purely unitary on the dimension equal to square of the graph order. The definition can be found in Quantum speed-up of Markov chain based algorithms by Szegedy. The definition of quantum search can be found in Direct Equivalence of Coined and Szegedy's Quantum Walks by Wong.

The abstract supertype is AbstractSzegedy with its default realization Szegedy. The model includes following types and methods:

QuantumWalk.AbstractSzegedy

QuantumWalk.QWEvolution

QuantumWalk.QWSearch

QuantumWalk.Szegedy

QuantumWalk.check_qwdynamics

QuantumWalk.check_qwdynamics

QuantumWalk.evolve

QuantumWalk.initial_state

QuantumWalk.measure

QuantumWalk.sqrtstochastic

Full docs

QuantumWalk.AbstractSzegedy — Type.

AbstractSzegedy

Abstract Szegedy model. Description of the default parameter can be found in https://arxiv.org/abs/1611.02238, where two oracle operator case is chosen. Default representation of AbstractSzegedy is Szegedy.

source

QuantumWalk.QWEvolution — Method.

QWEvolution(szegedy::AbstractSzegedy)

Create QWEvolution according to AbstractSzegedy model. By default, the constructed operator is of type SparseMatrixCSC.

source

QuantumWalk.QWSearch — Method.

QWSearch(szegedy::AbstractSzegedy, marked [, penalty])

Creates QWSearch according to AbstractSzegedy model. By default parameter penalty is set to 0. Evolution operators are constructed according to the definition from https://arxiv.org/abs/1611.02238.

source

QuantumWalk.Szegedy — Type.

Szegedy(graph[, stochastic, checkstochastic])

Default representation of AbstractSzegedy. Parameter graph needs to be a subtype of AbstractGraph from LightGraphs module, and parameter stochastic needs to be a Real column stochastic matrix. Flag checkstochastic decides about checking the stochastic properties.

Matrix stochastic defaults to the uniform walk operator, and checkstochastic deafults to false in case of default stochastic. If matrix stochastic is provided by the user, the default value of stochastic is true.

Matrix stochastic is changed into sqrtstochastic by taking element-wise square root.

source

QuantumWalk.check_qwdynamics — Method.

check_qwdynamics(szegedy::AbstractSzegedy, marked, parameters)

Check whetver combination of szegedy, marked and parameters produces valid QWSearch object. It checks where parameters consists of key :operators with corresponding value being list of SparseMatrixCSC. Furthermore operators needs to be square of size equals to square of graph(szegedy). order.

source

QuantumWalk.check_qwdynamics — Method.

check_qwdynamics(szegedy::AbstractSzegedy, marked, parameters)

Check whetver combination of szegedy, marked and parameters produces a valid QWEvolution object. It checks where parameters consists of key :operators with corresponding value being a list of SparseMatrixCSC objects. Furthermore operators need to be square of size equals to square of the order of graph(szegedy).

source

QuantumWalk.evolve — Method.

evolve(qwe::QWSearch{AbstractSzegedy}, state)

Multiplies state be each operator from operators from quatnum walk evolution qwe. Elements of operators and state should be of the same type.

julia> qss = QWSearch(Szegedy(CompleteGraph(4)), [1]);

julia> evolve(qss, initial_state(qss))
16-element SparseVector{Float64,Int64} with 12 stored entries:
  [2 ]  =  0.481125
  [3 ]  =  0.481125
  [4 ]  =  0.481125
  [5 ]  =  -0.288675
  [7 ]  =  -0.096225
  [8 ]  =  -0.096225
  [9 ]  =  -0.288675
  [10]  =  -0.096225
  [12]  =  -0.096225
  [13]  =  -0.288675
  [14]  =  -0.096225
  [15]  =  -0.096225

source

QuantumWalk.initial_state — Method.

initial_state(qss::QWSearch{<:AbstractSzegedy})

Generates typical initial state for Szegedy search, see https://arxiv.org/abs/1611.02238. Vectorizes and normalizes obtained sqrtstochastic matrix from model(qss). Note that the state may not be uniform supersposition in general, but it gives uniform superposition after partial tracing second system.

julia> qss = QWSearch(Szegedy(CompleteGraph(4)), [1]);

julia> initial_state(qss)
16-element SparseVector{Float64,Int64} with 12 stored entries:
  [2 ]  =  0.288675
  [3 ]  =  0.288675
  [4 ]  =  0.288675
  [5 ]  =  0.288675
  [7 ]  =  0.288675
  [8 ]  =  0.288675
  [9 ]  =  0.288675
  [10]  =  0.288675
  [12]  =  0.288675
  [13]  =  0.288675
  [14]  =  0.288675
  [15]  =  0.288675

source

QuantumWalk.measure — Method.

measure(qwd::AbstractSzegedy, state[, vertices])

Performes a measurement on state on vertices. vertices defaults to list of all vertices. It is defined as the measurement of partially traced on second system state https://arxiv.org/abs/1611.02238.

julia> qss = QWSearch(Szegedy(CompleteGraph(4)), [1]);

julia> state = evolve(qss, initial_state(qss));

julia> measure(qss, state)
4-element Array{Float64,1}:
 0.694444
 0.101852
 0.101852
 0.101852

julia> measure(qss, state, [1,3])
2-element Array{Float64,1}:
 0.694444
 0.101852

source

QuantumWalk.sqrtstochastic — Function.

sqrtstochastic(szegedy)

Return the sqrtstochastic element of szegedy. After element-wise squaring a column-stochastic matrix is obtained.

julia> szegedy = Szegedy(CompleteGraph(4));

julia> sqrtstochastic(szegedy)
4×4 SparseMatrixCSC{Float64,Int64} with 12 stored entries:
  [2, 1]  =  0.57735
  [3, 1]  =  0.57735
  [4, 1]  =  0.57735
  [1, 2]  =  0.57735
  [3, 2]  =  0.57735
  [4, 2]  =  0.57735
  [1, 3]  =  0.57735
  [2, 3]  =  0.57735
  [4, 3]  =  0.57735
  [1, 4]  =  0.57735
  [2, 4]  =  0.57735
  [3, 4]  =  0.57735

julia> full(sqrtstochastic(szegedy).^2)
4×4 Array{Float64,2}:
 0.0       0.333333  0.333333  0.333333
 0.333333  0.0       0.333333  0.333333
 0.333333  0.333333  0.0       0.333333
 0.333333  0.333333  0.333333  0.0

source